adding gitignore file
