+++ src/bicycles/rides/BikeRide.java
@@ -1,4 +1,4 @@
-package bicycles.Rides;
+package bicycles.rides;
 
 
   public interface BikeRide {
