+++ tests/bicycles/specifications/RoadBikeSpec.java
@@ -2,8 +2,6 @@ package bicycles.specifications;
 
 import bicycles.Bicycle;
 import bicycles.BicycleType;
-import bicycles.Rides.BikeRide;
-import bicycles.Rides.BikeRideOne;
 import org.junit.jupiter.api.Test;
 
 public class RoadBikeSpec {
@@ -13,11 +11,9 @@ public class RoadBikeSpec {
 
         BicycleSpecification roadBikeSpec = new BicycleSpecification(11, 4, BicycleType.RoadBike);
         Bicycle roadBikespecs = new BicycleFromSpec(roadBikeSpec);
-        BikeRide rRide = new BikeRideOne(roadBikespecs);
-        rRide.ride();
+
 
         System.out.println(roadBikeSpec.getAccelerationSpeed());
-
         System.out.println(roadBikeSpec.getBrakeSpeed());
     }
 }
