+++ tests/bicycles/specifications/TandemSpec.java
@@ -2,8 +2,6 @@ package bicycles.specifications;
 
 import bicycles.Bicycle;
 import bicycles.BicycleType;
-import bicycles.Rides.BikeRide;
-import bicycles.Rides.BikeRideOne;
 import org.junit.jupiter.api.Test;
 
 public class TandemSpec {
@@ -13,8 +11,6 @@ public class TandemSpec {
 
         BicycleSpecification tandemspec = new BicycleSpecification(12, 7, BicycleType.Tandem);
         Bicycle Tandemspecs = new BicycleFromSpec(tandemspec);
-        BikeRide tSpecs = new BikeRideOne(Tandemspecs);
-        tSpecs.ride();
 
         System.out.println(tandemspec.getAccelerationSpeed());
